requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
groq==0.11.0
httpx==0.27.2
pytz==2024.2
python-dotenv==1.0.1
